# Conflicts:
#	acm-services/acm-service-config/src/main/resources/spring/spring-library-config.xml


Former-commit-id: ace8079dc2d9dcd47ff0a9611d772d2bf45edc25

+++ acm-plugins/acm-default-plugins/acm-case-file-plugin/src/main/java/com/armedia/acm/plugins/casefile/service/CaseFileEventListener.java
@@ -60,50 +60,53 @@ public class CaseFileEventListener implements ApplicationListener<AcmObjectHisto
                     String json = acmObjectHistoryExisting.getObjectString();
                     CaseFile existing = converter.unmarshall(json, CaseFile.class);
 
-                    acmAssignment.setOldAssignee(ParticipantUtils.getAssigneeIdFromParticipants(existing.getParticipants()));
-
-                    if (isPriorityChanged(existing, updatedCaseFile))
+                    if (existing != null)
                     {
-                        getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "priority.changed");
-                    }
+                        acmAssignment.setOldAssignee(ParticipantUtils.getAssigneeIdFromParticipants(existing.getParticipants()));
 
-                    if (isDetailsChanged(existing, updatedCaseFile))
-                    {
-                        getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "details.changed");
-                    }
-
-                    String title = existing.getTitle();
-                    String updatedTitle = updatedCaseFile.getTitle();
-                    if (!Objects.equals(title, updatedTitle))
-                    {
-                        getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "title.changed",
-                                "Case File Title changed from " + title + " to " + updatedTitle);
-                    }
-
-
-                    String owningGroup = ParticipantUtils.getOwningGroupIdFromParticipants(existing.getParticipants());
-                    String updatedOwningGroup = ParticipantUtils.getOwningGroupIdFromParticipants(updatedCaseFile.getParticipants());
-                    if (!Objects.equals(owningGroup, updatedOwningGroup))
-                    {
-                        AcmParticipant updatedParticipant = updatedCaseFile.getParticipants().stream().filter(p -> "owning group".equals(p.getParticipantType())).findFirst().orElse(null);
-                        getCaseFileEventUtility().raiseParticipantsModifiedInCaseFile(updatedParticipant, updatedCaseFile, event.getIpAddress(), "changed",
-                                "Owning Group Changed from " + owningGroup + " to " + updatedOwningGroup);
-                    }
-
-                    checkParticipants(existing, updatedCaseFile, event.getIpAddress());
-
-                    if (isStatusChanged(existing, updatedCaseFile))
-                    {
-                        String calId = updatedCaseFile.getContainer().getCalendarFolderId();
-                        if (Objects.equals(updatedCaseFile.getStatus(), caseFileStatusClosed)
-                                && shouldDeleteCalendarFolder && calId != null)
+                        if (isPriorityChanged(existing, updatedCaseFile))
                         {
-
-                            //delete shared calendar if case closed
-                            getCalendarService().deleteFolder(updatedCaseFile.getContainer().getContainerObjectId(),
-                                    calId, DeleteMode.MoveToDeletedItems);
+                            getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "priority.changed");
                         }
-                        getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "status.changed");
+
+                        if (isDetailsChanged(existing, updatedCaseFile))
+                        {
+                            getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "details.changed");
+                        }
+
+                        String title = existing.getTitle();
+                        String updatedTitle = updatedCaseFile.getTitle();
+                        if (!Objects.equals(title, updatedTitle))
+                        {
+                            getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "title.changed",
+                                    "Case File Title changed from " + title + " to " + updatedTitle);
+                        }
+
+
+                        String owningGroup = ParticipantUtils.getOwningGroupIdFromParticipants(existing.getParticipants());
+                        String updatedOwningGroup = ParticipantUtils.getOwningGroupIdFromParticipants(updatedCaseFile.getParticipants());
+                        if (!Objects.equals(owningGroup, updatedOwningGroup))
+                        {
+                            AcmParticipant updatedParticipant = updatedCaseFile.getParticipants().stream().filter(p -> "owning group".equals(p.getParticipantType())).findFirst().orElse(null);
+                            getCaseFileEventUtility().raiseParticipantsModifiedInCaseFile(updatedParticipant, updatedCaseFile, event.getIpAddress(), "changed",
+                                    "Owning Group Changed from " + owningGroup + " to " + updatedOwningGroup);
+                        }
+
+                        checkParticipants(existing, updatedCaseFile, event.getIpAddress());
+
+                        if (isStatusChanged(existing, updatedCaseFile))
+                        {
+                            String calId = updatedCaseFile.getContainer().getCalendarFolderId();
+                            if (Objects.equals(updatedCaseFile.getStatus(), caseFileStatusClosed)
+                                    && shouldDeleteCalendarFolder && calId != null)
+                            {
+
+                                //delete shared calendar if case closed
+                                getCalendarService().deleteFolder(updatedCaseFile.getContainer().getContainerObjectId(),
+                                        calId, DeleteMode.MoveToDeletedItems);
+                            }
+                            getCaseFileEventUtility().raiseCaseFileModifiedEvent(updatedCaseFile, event.getIpAddress(), "status.changed");
+                        }
                     }
                 }
